mostly moving files around + fix the sporadic job cancelling
