+++ .gitignore
@@ -7,6 +7,8 @@ data*
 .idea
 .vscode
 
+# FIXME: tmporary irnogre, but test folder should be linted, fix issues
+test/*
 /test/.test-ca.pem
 /test/.test-ca.key
