--- .pre-commit-config.yaml
@@ -1,27 +1,0 @@
-repos:
-  - repo: https://github.com/pre-commit/pre-commit-hooks
-    rev: v2.4.0
-    hooks:
-      - id: trailing-whitespace
-      - id: end-of-file-fixer
-  - repo: local
-    hooks:
-      - id: eslint
-        name: eslint
-        entry: node_modules/.bin/eslint
-        language: node
-        files: \.[^\.]*(t|j)s$
-        exclude: ".env.cjs|test\/.*"
-        args: ["--quiet"]
-      - id: prettier
-        name: prettier
-        entry: node_modules/.bin/prettier
-        language: node
-        files: \.(html|json)$
-        args: ["--check"]
-      - id: typecheck
-        name: typecheck
-        entry: node_modules/.bin/tsc-files
-        language: node
-        files: \.ts$
-        args: ["--noEmit"]
