+++ package.json
@@ -1,6 +1,6 @@
 {
-  "name": "try-runtime-bot",
-  "description": "try-runtime-bot",
+  "name": "command-bot",
+  "description": "command-bot",
   "version": "0.0.0",
   "author": "Parity <admin@parity.io> (https://parity.io)",
   "license": "Apache-2.0",
@@ -64,10 +64,11 @@
     "pretest:integration": "yarn build",
     "test:integration": "node --experimental-vm-modules node_modules/.bin/jest --runInBand -c jest.integration.config.cjs",
     "test": "yarn test:unit && yarn test:integration",
-    "typecheck": "tsc --noEmit",
-    "prettier": "prettier --loglevel silent '{*,**/*}.{json,html}'",
     "prelint": "yarn schema",
-    "lint": "pre-commit run --color=always --all",
+    "lint": "yarn eslint && yarn typecheck && yarn prettier",
+    "typecheck": "tsc --noEmit",
+    "prettier": "prettier --check --loglevel silent '{*,**/*}.{json,html}'",
+    "eslint": "eslint --quiet '{*,**/*}.{cjs,ts}'",
     "fix:eslint": "eslint --fix",
     "fix:prettier": "prettier --write",
     "fix": "yarn fix:eslint '{*,**/*}.{cjs,ts}' && yarn fix:prettier '{*,**/*}.json'",
